# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "运维管理平台 API 文档",
        "title": "OpsHub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/kubernetes/clusters/kubeconfig/existing": {
            "get": {
                "description": "获取当前用户在指定集群的最新KubeConfig",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes/Cluster"
                ],
                "summary": "获取用户现有的KubeConfig",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "clusterId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/kubernetes/clusters/{id}": {
            "get": {
                "description": "获取 Kubernetes 集群详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes/Cluster"
                ],
                "summary": "获取集群详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "更新 Kubernetes 集群信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes/Cluster"
                ],
                "summary": "更新集群",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除 Kubernetes 集群",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes/Cluster"
                ],
                "summary": "删除集群",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/kubernetes/clusters/{id}/test": {
            "post": {
                "description": "测试 Kubernetes 集群连接状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes/Cluster"
                ],
                "summary": "测试集群连接",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/plugins/kubernetes/clusters/{id}/config": {
            "get": {
                "description": "获取集群的 KubeConfig 配置（已解密）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes/Cluster"
                ],
                "summary": "获取集群凭证",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/plugins/kubernetes/clusters/{id}/sync": {
            "post": {
                "description": "同步指定集群的状态信息（节点数、Pod数等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kubernetes/Cluster"
                ],
                "summary": "同步集群状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "集群ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}